+++ Application/RDD.Application/Controllers/AppController.cs
@@ -1,6 +1,5 @@
 ﻿using RDD.Domain;
 using RDD.Domain.Models.Querying;
-using RDD.Infra;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;